pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "maat"
version = "0.1.0"
description = "LLM chess architecture experiments"
readme = "implementation_plan.md"
requires-python = ">=3.11"
dependencies = [
    "python-chess>=1.999",
    "pydantic>=2.6",
    "langgraph>=0.2.0",
    "google-genai>=0.7.0",
    "langchain-google-genai>=2.1.0",
    "langchain-core>=0.3.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0",
    "stockfish>=3.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.6.0",
    "black>=24.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.black]
line-length = 100
target-version = ["py311"]
